Added bucket.GetAny and cursor.IsBucket functions
GetAny retrieves a value or nested bucket for a key in the bucket.
Returns ok if the key exists and a value or a nested bucket.
The returned value or bucket instance is only valid for the life of
the transaction.

    func (b *Bucket) GetAny(key []byte) (
	value []byte, bucket *Bucket, ok bool,
    )

IsBucket returns true if the current key under the cursor is a
bucket.

    func (c *Cursor) IsBucket() bool {
	return c.cisb
    }

For example:

    c := b.Cursor()
    for k, v := c.First(); k != nil; k, v = c.Next() {
        if c.IsBucket(){
    	// current key is a nested bucket
        } else {
    	// current key is a []byte value
        }
    }